requirement.txt
opencv-python
numpy
pyzbar
streamlit
Pillow